Conflicts:

	fbreader/data/default/keymap.desktop.xml
	fbreader/data/default/styles.desktop.xml
	fbreader/src/fbreader/BookTextView.h
	fbreader/src/fbreader/FBReader.cpp
	fbreader/src/fbreader/FBReader.h
	fbreader/src/fbreader/FBReaderActionCode.cpp
	fbreader/src/fbreader/FBReaderActions.cpp
	fbreader/src/fbreader/FBReaderActions.h
	fbreader/src/fbreader/FBView.cpp
	fbreader/src/fbreader/SearchActions.cpp
	fbreader/src/formats/fb2/FB2BookReader.cpp
	makefiles/arch/desktop.mk
	makefiles/target.mk
	zlibrary/core/src/unix/xmlconfig/XMLConfigIO.cpp
	zlibrary/core/src/view/ZLView.h
	zlibrary/text/src/view/ZLTextView_paint.cpp
	zlibrary/ui/Makefile

+++ fbreader/src/formats/chm/HHCReader.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004-2008 Geometer Plus <contact@geometerplus.com>
+ * Copyright (C) 2004-2009 Geometer Plus <contact@geometerplus.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -96,7 +96,7 @@ bool HHCReader::tagHandler(const HtmlTag &tag) {
 	return true;
 }
 
-bool HHCReader::characterDataHandler(const char*, int, bool) {
+bool HHCReader::characterDataHandler(const char*, size_t, bool) {
 	return true;
 }
